extract `getUpdatedFields`
