+++ shared-libs/cht-datasource/src/local/libs/core.ts
@@ -16,3 +16,24 @@ export const normalizeFreetext = (
 ): string => {
   return freetext.trim().toLowerCase();
 };
+
+/** @internal*/
+export const getUpdatedFields = (
+  originalDoc:Record<string, unknown>,
+  updatedDoc : Record<string, unknown>,
+  ignoreUpdateFields : Set<string>
+):Record<string, unknown> => {
+  const updatedFields:Record<string, unknown> = {};
+  for (const key of Object.keys(originalDoc)) {
+    if (ignoreUpdateFields.has(key)) {
+      continue;
+    }
+    if (updatedDoc[key] ===undefined) {
+      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+      delete originalDoc[key];
+      continue;
+    }
+    updatedFields[key] = updatedDoc[key];
+  }
+  return updatedFields;
+};
